refactor(tasks): extract routes + schemas/guard/search-doc from index
WS6 plugin god-file split, following the audit's seam map. The tasks plugin's
declarative route array and the helpers it shares with the exec tools move out
of the 1,089-line index.ts:

- lib/routes.ts — the full declarative route array (board read/summary/details/
  run-history + create/update/delete/move/assign/log/block/dependency/reorder/
  complete + the search route). Handlers are ctx-parameterized; indexTask and
  taskEditGuard take ctx as a param, so no plugin-scope wiring is needed.
- lib/task-schemas.ts — the canonical COLUMNS tuple + every zod request/response
  schema, shared by routes and exec tools.
- lib/edit-guard.ts — taskEditGuard + guardResponse (optimistic versioning +
  freeze-on-complete). Covered by tests/plugins/tasks/task-edit-safety.test.ts.
- lib/search-doc.ts — taskToSearchDoc + indexTask (precedent: assets/lib/search-doc).

index.ts: 1,089 -> 539 lines. Pure relocation — definePlugin now does
`routes: tasksRoutes`, and the exec-tool blocks (still in activate) import the
shared helpers. The optional behavioral "redesign opportunities" the audit lists
(REST/MCP guard inconsistency, identifier-fallback dedup, dead .catch noise) are
deliberately NOT taken here — they're behavior-touching, noted for follow-up.

Deferred to a follow-up: the ~370-line exec-tools block -> lib/exec-tools.ts and
lib/maintenance.ts, which slim activate() to the thin-index shape.

Verified: typecheck + lint clean; tasks suite 206/0 (incl. routes + edit-safety);
full suite 5123/0; binary build; isolated-home boot smoke (Tasks activates; GET
/, /summary, and POST create all 200 through the relocated handlers).
