+++ plugins/tasks/index.ts
@@ -4,13 +4,12 @@
  */
 import { z } from 'zod'
 import type { BakinPlugin, PluginContext } from '@bakin/core/plugin-types'
-import { definePlugin, defineRoute, searchRoute, type PluginContextLite } from '@bakin/core/routing'
+import { definePlugin } from '@bakin/core/routing'
 import {
   readTaskboard,
   deleteTask,
   assignTask,
   updateTask,
-  reorderTasks,
   archiveOldTasks,
   autoArchiveDoneTasks,
   getTask,
@@ -26,8 +25,7 @@ import {
   triggerDispatch,
 } from '../../src/core/task-service'
 import { createLogger } from '../../src/core/logger'
-import { getLiveRun, hasCompletion } from '../../src/core/execution-ledger'
-import { readTaskOutcome, readTaskRuns } from './lib/runs-reader'
+import { getLiveRun } from '../../src/core/execution-ledger'
 import { getContentDir } from '../../packages/core/src/content-dir'
 import {
   checkSessionDeathIncidents,
@@ -39,6 +37,10 @@ import {
 } from './lib/health-checks'
 import { queryAuditEvents } from '../../src/core/audit'
 import type { Task, ColumnId } from './types'
+import { tasksRoutes } from './lib/routes'
+import { taskEditGuard } from './lib/edit-guard'
+import { indexTask, taskToSearchDoc } from './lib/search-doc'
+import { COLUMNS } from './lib/task-schemas'
 
 const log = createLogger('tasks')
 
@@ -59,7 +61,6 @@ function buildChecklistNudge(description: string | undefined): string | undefine
   return `This task enumerates ${enumeratedLines.length} items. Consider formatting deliverables as a markdown checklist ("- [ ] …") — agents then produce and save each one in succession (write file → assets_save → log) instead of drafting everything in one oversized response.`
 }
 
-const COLUMNS = ['backlog', 'todo', 'inProgress', 'review', 'done', 'blocked', 'archived'] as const
 let maintenanceTimers: Array<ReturnType<typeof setInterval>> = []
 
 function clearMaintenanceTimers(): void {
@@ -69,564 +70,13 @@ function clearMaintenanceTimers(): void {
   maintenanceTimers = []
 }
 
-// ─── Search-document helpers (module scope) ──────────────────────────────
-
-function taskToSearchDoc(task: Task, column: ColumnId): Record<string, unknown> {
-  const logText = task.log?.map(l => `[${l.timestamp} ${l.author}] ${l.message}`).join('\n') || ''
-  return {
-    title: task.title,
-    description: task.description || '',
-    agent: task.agent || '',
-    created_by: task.createdBy || '',
-    status: column,
-    project_id: task.projectId || '',
-    workflow_id: task.workflowId || '',
-    source_plugin_id: task.source?.pluginId || '',
-    source_entity_type: task.source?.entityType || '',
-    source_entity_id: task.source?.entityId || '',
-    source_purpose: task.source?.purpose || '',
-    available_at: task.availableAt || '',
-    due_at: task.dueAt || '',
-    log_text: logText,
-    blocked_reason: task.blockedReason || '',
-    updated_at: new Date().toISOString(),
-  }
-}
-
-async function indexTask(ctx: PluginContextLite, taskId: string): Promise<void> {
-  try {
-    const board = readTaskboard()
-    const columns = board.columns as unknown as Record<string, Task[]>
-    for (const [colName, tasks] of Object.entries(columns)) {
-      const task = tasks.find(t => t.id === taskId)
-      if (task) {
-        await ctx.search.index(taskId, taskToSearchDoc(task, colName as ColumnId))
-        return
-      }
-    }
-  } catch (err) {
-    log.warn('Failed to index task', { taskId, error: err instanceof Error ? err.message : String(err) })
-  }
-}
-
-// ─── Schemas (module scope) ──────────────────────────────────────────────
-
-const taskIdParams = z.object({ taskId: z.string() })
-
-const runsHistoryQuery = z.object({ limit: z.coerce.number().int().positive().optional() })
-
-const okResponse = z.object({ ok: z.literal(true) })
-const errorResponse = z.object({ error: z.string() })
-
-const taskBoardResponse = z.object({}).passthrough()  // structural; defined elsewhere
-
-const taskSummaryResponse = z.object({ blocked: z.number(), review: z.number() })
-
-const taskSourceSchema = z.object({
-  pluginId: z.string().optional(),
-  entityType: z.string().optional(),
-  entityId: z.string().optional(),
-  purpose: z.string().optional(),
-})
-
-const createTaskBody = z.object({
-  id: z.string().optional(),
-  title: z.string().min(1),
-  description: z.string().optional(),
-  column: z.enum(COLUMNS).optional(),
-  assignee: z.string().optional(),
-  workflowId: z.string().optional(),
-  skipWorkflowReason: z.string().optional(),
-  createdBy: z.string().optional(),
-  parentId: z.string().optional(),
-  projectId: z.string().optional(),
-  availableAt: z.string().optional(),
-  dueAt: z.string().optional(),
-  source: taskSourceSchema.optional(),
-})
-const createTaskResponse = z.object({
-  ok: z.literal(true),
-  id: z.string(),
-  workflowId: z.string().optional(),
-  suggestedWorkflow: z.string().optional(),
-})
-
-const updateTaskBody = z.object({
-  id: z.string().optional(),
-  originalTitle: z.string().optional(),
-  title: z.string().optional(),
-  description: z.string().optional(),
-  agent: z.string().optional(),
-  column: z.string().optional(),
-  workflowId: z.string().optional(),
-  availableAt: z.string().nullable().optional(),
-  dueAt: z.string().nullable().optional(),
-  source: taskSourceSchema.nullable().optional(),
-  /** Optimistic concurrency: refuse with 409 when stale. */
-  expectedVersion: z.number().optional(),
-})
-
-const deleteTaskBody = z.object({
-  id: z.string().optional(),
-  title: z.string().optional(),
-}).optional()
-
-const moveTaskBody = z.object({
-  id: z.string().optional(),
-  title: z.string().optional(),
-  from: z.string().optional(),
-  to: z.string(),
-  agent: z.string(),
-  reason: z.string().optional(),
-  channel: z.string().optional(),
-})
-
-const assignTaskBody = z.object({
-  id: z.string().optional(),
-  title: z.string().optional(),
-  agent: z.string().optional(),
-})
-
-const logEntryBody = z.object({
-  id: z.string().optional(),
-  title: z.string().optional(),
-  author: z.string().optional(),
-  agent: z.string().optional(),
-  message: z.string().min(1),
-})
-
-const blockTaskBody = z.object({
-  id: z.string().optional(),
-  title: z.string().optional(),
-  agent: z.string().optional(),
-  reason: z.string().min(1),
-})
-
-const dependencyBody = z.object({
-  id: z.string().optional(),
-  dependsOn: z.string().min(1),
-})
-
-const reorderBody = z.object({
-  columnId: z.string().min(1),
-  orderedIds: z.array(z.string()),
-})
-
-const completeTaskBody = z.object({
-  id: z.string().optional(),
-  agent: z.string().optional(),
-  summary: z.string().optional(),
-})
-
-// ─── Edit-safety guard (optimistic versioning + freeze-on-complete) ───────
-//
-// Content mutations (update/assign/dependency/block) pass through here.
-// - A completed task (completions row in the execution ledger) refuses
-//   mutation until explicitly reopened (moved out of Done). Move/complete
-//   routes are exempt — they ARE the reopen/complete paths.
-// - A stale expectedVersion gets a 409 + edit_conflict audit so contention
-//   is measurable; omitting expectedVersion stays last-write-wins.
-function taskEditGuard(
-  ctx: { activity: { audit: (event: string, agent: string, data?: Record<string, unknown>) => void } },
-  identifier: string,
-  opts: { agent?: string; expectedVersion?: number } = {},
-): { status: number; error: string; currentVersion?: number } | null {
-  const task = getTask(identifier)
-  if (!task) return null // unknown ids fall through to the handler's own error path
-  if (hasCompletion(task.id)) {
-    return { status: 409, error: `Task ${task.id} is completed — reopen it (move it out of Done) before editing.` }
-  }
-  const currentVersion = task.version ?? 0
-  if (opts.expectedVersion !== undefined && currentVersion !== opts.expectedVersion) {
-    ctx.activity.audit('edit_conflict', opts.agent ?? 'system', {
-      taskId: task.id,
-      expectedVersion: opts.expectedVersion,
-      currentVersion,
-    })
-    return {
-      status: 409,
-      error: `Version conflict on ${task.id}: expected version ${opts.expectedVersion}, current is ${currentVersion}. Refetch and retry.`,
-      currentVersion,
-    }
-  }
-  return null
-}
-
-function guardResponse(guard: { status: number; error: string; currentVersion?: number }): Response {
-  return Response.json(
-    { error: guard.error, ...(guard.currentVersion !== undefined ? { currentVersion: guard.currentVersion } : {}) },
-    { status: guard.status },
-  )
-}
-
-// ─── Routes (declarative) ────────────────────────────────────────────────
-
-const routes = [
-  defineRoute({
-    path: '/',
-    method: 'GET',
-    summary: 'List all tasks',
-    description: 'Returns the full kanban board state, grouped by column.',
-    responses: { 200: taskBoardResponse, 500: errorResponse },
-    handler: async () => {
-      try {
-        const board = await readTaskboard()
-        return Response.json(board)
-      } catch (err) {
-        return Response.json({ error: err instanceof Error ? err.message : String(err) }, { status: 500 })
-      }
-    },
-  }),
-
-  defineRoute({
-    path: '/summary',
-    method: 'GET',
-    summary: 'Counts of tasks needing attention (nav-badge source)',
-    description: 'Returns blocked/review counts only — cheap source for the Tasks nav badge.',
-    responses: { 200: taskSummaryResponse, 500: errorResponse },
-    handler: async () => {
-      try {
-        const board = await readTaskboard()
-        return Response.json({
-          blocked: board.columns.blocked.length,
-          review: board.columns.review.length,
-        })
-      } catch (err) {
-        return Response.json({ error: err instanceof Error ? err.message : String(err) }, { status: 500 })
-      }
-    },
-  }),
-
-  defineRoute({
-    path: '/:taskId',
-    method: 'GET',
-    summary: 'Get a task by ID',
-    params: taskIdParams,
-    responses: { 200: z.object({}).passthrough(), 404: errorResponse, 500: errorResponse },
-    handler: async (_req, _ctx, { params }) => {
-      const result = await getTaskDetails(params.taskId)
-      if (!result) {
-        return Response.json({ error: 'Task not found' }, { status: 404 })
-      }
-      return Response.json(result)
-    },
-  }),
-
-  defineRoute({
-    path: '/:taskId/runs',
-    method: 'GET',
-    summary: 'Get dispatch run history for a task',
-    params: taskIdParams,
-    query: runsHistoryQuery,
-    responses: { 200: z.object({}).passthrough() },
-    handler: async (_req, _ctx, { params, query }) => {
-      // Unknown task → empty list, never an error (a not-yet-dispatched task has no runs).
-      // Column fallback is gated on dispatch history so unknown ids can't trigger
-      // the task store's full shard walk on every request.
-      const runs = readTaskRuns(params.taskId, query.limit ?? 50)
-      return Response.json({ runs, outcome: readTaskOutcome(params.taskId, runs.length > 0) })
-    },
-  }),
-
-  defineRoute({
-    path: '/',
-    method: 'POST',
-    summary: 'Create a task',
-    description: 'Creates a task on the kanban board. Auto-matches a workflow by title when workflowId is omitted.',
-    body: createTaskBody,
-    responses: { 200: createTaskResponse, 500: errorResponse },
-    handler: async (_req, ctx, { body }) => {
-      try {
-        const result = await createTaskWithEffects({
-          id: body.id,
-          title: body.title,
-          column: body.column as ColumnId | undefined,
-          assignee: body.assignee,
-          description: body.description,
-          workflowId: body.workflowId,
-          skipWorkflowReason: body.skipWorkflowReason,
-          createdBy: body.createdBy || 'system',
-          parentId: body.parentId,
-          projectId: body.projectId,
-          availableAt: body.availableAt,
-          dueAt: body.dueAt,
-          source: body.source,
-          channel: 'rest',
-        })
-        ctx.activity.log(body.createdBy || 'system', `Created task "${body.title}"`, { taskId: result.id })
-        indexTask(ctx, result.id).catch(() => {})
-        return Response.json({ ok: true as const, id: result.id, workflowId: result.workflowId, suggestedWorkflow: result.suggestedWorkflow })
-      } catch (err) {
-        return Response.json({ error: err instanceof Error ? err.message : String(err) }, { status: 500 })
-      }
-    },
-  }),
-
-  defineRoute({
-    path: '/:taskId',
-    method: 'PUT',
-    summary: 'Update a task',
-    params: taskIdParams,
-    body: updateTaskBody,
-    responses: { 200: okResponse, 400: errorResponse, 409: errorResponse, 500: errorResponse },
-    handler: async (_req, ctx, { params, body }) => {
-      const identifier = params.taskId || body.id || body.originalTitle
-      if (!identifier) {
-        return Response.json({ error: 'taskId required' }, { status: 400 })
-      }
-      const guard = taskEditGuard(ctx, identifier, { agent: body.agent, expectedVersion: body.expectedVersion })
-      if (guard) return guardResponse(guard)
-      try {
-        await updateTask(identifier, {
-          title: body.title,
-          description: body.description,
-          agent: body.agent,
-          column: body.column as ColumnId | undefined,
-          workflowId: body.workflowId,
-          availableAt: body.availableAt,
-          dueAt: body.dueAt,
-          source: body.source,
-        })
-        const agent = body.agent || 'system'
-        ctx.activity.audit('updated', agent, { taskId: identifier })
-        ctx.activity.log(agent, `Updated task "${identifier}"`, { taskId: identifier })
-        indexTask(ctx, identifier).catch(() => {})
-        return Response.json({ ok: true as const })
-      } catch (err) {
-        return Response.json({ error: err instanceof Error ? err.message : String(err) }, { status: 500 })
-      }
-    },
-  }),
-
-  defineRoute({
-    path: '/:taskId',
-    method: 'DELETE',
-    summary: 'Delete a task',
-    params: taskIdParams,
-    body: { contentType: '*/*', schema: deleteTaskBody },  // optional body fallback (id/title)
-    responses: { 200: okResponse, 400: errorResponse, 500: errorResponse },
-    handler: async (req, ctx, { params }) => {
-      let body: any = {}
-      try { body = await req.clone().json() } catch { /* no body is fine */ }
-      const identifier = params.taskId || body.id || body.title
-      if (!identifier) {
-        return Response.json({ error: 'taskId required' }, { status: 400 })
-      }
-      try {
-        await deleteTask(identifier as string)
-        ctx.activity.audit('deleted', 'system', { taskId: identifier })
-        ctx.activity.log('system', `Deleted task "${identifier}"`, { taskId: identifier as string })
-        ctx.search.remove(identifier as string).catch(() => {})
-        return Response.json({ ok: true as const })
-      } catch (err) {
-        return Response.json({ error: err instanceof Error ? err.message : String(err) }, { status: 500 })
-      }
-    },
-  }),
-
-  defineRoute({
-    path: '/:taskId/move',
-    method: 'POST',
-    summary: 'Move a task to a different column',
-    params: taskIdParams,
-    body: moveTaskBody,
-    responses: { 200: okResponse, 400: errorResponse, 403: errorResponse, 409: errorResponse, 500: errorResponse },
-    handler: async (_req, ctx, { params, body }) => {
-      const identifier = params.taskId || body.id || body.title
-      if (!identifier) {
-        return Response.json({ error: 'taskId and to required' }, { status: 400 })
-      }
-      const { from, to, agent, reason } = body
-      if (to === 'blocked') {
-        if (!reason) {
-          return Response.json({ error: 'reason required when moving to blocked' }, { status: 400 })
-        }
-        try {
-          const { alreadyComplete } = await blockTaskWithEffects(identifier, reason, agent, (body.channel === 'human' && agent === 'human') ? 'human' : 'rest')
-          if (alreadyComplete) {
-            return Response.json({ error: `Task ${identifier} is completed — reopen it (move it out of Done) before blocking.` }, { status: 409 })
-          }
-          ctx.activity.log(agent, `Blocked task: ${reason}`, { taskId: identifier })
-          indexTask(ctx, identifier).catch(() => {})
-          return Response.json({ ok: true as const })
-        } catch (err) {
-          return Response.json({ error: (err as Error).message }, { status: 500 })
-        }
-      }
-      try {
-        const effectiveChannel = (body.channel === 'human' && agent === 'human') ? 'human' as const : 'rest' as const
-        const { alreadyComplete } = await moveTaskWithEffects(identifier, to, agent, { from, channel: effectiveChannel })
-        if (!alreadyComplete) {
-          ctx.activity.log(agent, `Moved task to "${to}"`, { taskId: identifier })
-          indexTask(ctx, identifier).catch(() => {})
-        }
-        return Response.json({ ok: true as const, ...(alreadyComplete ? { alreadyComplete: true as const } : {}) })
-      } catch (err) {
-        const msg = (err as Error).message
-        if (msg.includes('Workflow tasks cannot be moved')) {
-          return Response.json({ error: msg }, { status: 403 })
-        }
-        return Response.json({ error: msg }, { status: 500 })
-      }
-    },
-  }),
-
-  defineRoute({
-    path: '/:taskId/assign',
-    method: 'POST',
-    summary: 'Assign a task to an agent',
-    params: taskIdParams,
-    body: assignTaskBody,
-    responses: { 200: okResponse, 400: errorResponse, 409: errorResponse, 500: errorResponse },
-    handler: async (_req, ctx, { params, body }) => {
-      const identifier = params.taskId || body.id || body.title
-      if (!identifier) {
-        return Response.json({ error: 'taskId required' }, { status: 400 })
-      }
-      const guard = taskEditGuard(ctx, identifier, { agent: body.agent })
-      if (guard) return guardResponse(guard)
-      try {
-        await assignTask(identifier, body.agent || '')
-        ctx.activity.audit('assigned', 'system', { taskId: identifier, agent: body.agent || '' })
-        ctx.activity.log('system', `Assigned task to "${body.agent || 'unassigned'}"`, { taskId: identifier })
-        ctx.search.transform(identifier, [{ op: '$set', field: 'agent', value: body.agent || '' }]).catch(() => {})
-        return Response.json({ ok: true as const })
-      } catch (err) {
-        return Response.json({ error: err instanceof Error ? err.message : String(err) }, { status: 500 })
-      }
-    },
-  }),
-
-  defineRoute({
-    path: '/:taskId/log',
-    method: 'POST',
-    summary: 'Add a log entry to a task',
-    params: taskIdParams,
-    body: logEntryBody,
-    responses: { 200: okResponse, 400: errorResponse, 500: errorResponse },
-    handler: async (_req, ctx, { params, body }) => {
-      const identifier = params.taskId || body.id || body.title
-      if (!identifier) {
-        return Response.json({ error: 'taskId required' }, { status: 400 })
-      }
-      try {
-        await logProgress(identifier, body.author || 'system', body.message, 'rest')
-        ctx.activity.log(body.agent || body.author || 'system', `Logged progress on task ${identifier}`, { taskId: identifier })
-        return Response.json({ ok: true as const })
-      } catch (err) {
-        return Response.json({ error: err instanceof Error ? err.message : String(err) }, { status: 500 })
-      }
-    },
-  }),
-
-  defineRoute({
-    path: '/:taskId/block',
-    method: 'POST',
-    summary: 'Mark a task as blocked',
-    params: taskIdParams,
-    body: blockTaskBody,
-    responses: { 200: okResponse, 400: errorResponse, 409: errorResponse, 500: errorResponse },
-    handler: async (_req, ctx, { params, body }) => {
-      const identifier = params.taskId || body.id || body.title
-      if (!identifier) {
-        return Response.json({ error: 'taskId required' }, { status: 400 })
-      }
-      const guard = taskEditGuard(ctx, identifier, { agent: body.agent })
-      if (guard) return guardResponse(guard)
-      try {
-        await blockTaskWithEffects(identifier, body.reason, body.agent || 'system', 'rest')
-        ctx.activity.log(body.agent || 'system', `Blocked task: ${body.reason}`, { taskId: identifier })
-        indexTask(ctx, identifier).catch(() => {})
-        return Response.json({ ok: true as const })
-      } catch (err) {
-        return Response.json({ error: err instanceof Error ? err.message : String(err) }, { status: 500 })
-      }
-    },
-  }),
-
-  defineRoute({
-    path: '/:taskId/dependency',
-    method: 'POST',
-    summary: 'Set a dependency between tasks',
-    params: taskIdParams,
-    body: dependencyBody,
-    responses: { 200: okResponse, 400: errorResponse, 409: errorResponse, 500: errorResponse },
-    handler: async (_req, ctx, { params, body }) => {
-      const taskId = params.taskId || body.id
-      if (!taskId) {
-        return Response.json({ error: 'taskId required' }, { status: 400 })
-      }
-      const guard = taskEditGuard(ctx, taskId)
-      if (guard) return guardResponse(guard)
-      try {
-        await setDependencyWithEffects(taskId, body.dependsOn, 'rest')
-        ctx.activity.log('system', `Set dependency on ${body.dependsOn}`, { taskId })
-        return Response.json({ ok: true as const })
-      } catch (err) {
-        return Response.json({ error: err instanceof Error ? err.message : String(err) }, { status: 500 })
-      }
-    },
-  }),
-
-  defineRoute({
-    path: '/reorder',
-    method: 'POST',
-    summary: 'Reorder tasks within a column',
-    body: reorderBody,
-    responses: { 200: okResponse, 400: errorResponse, 500: errorResponse },
-    handler: async (_req, ctx, { body }) => {
-      try {
-        await reorderTasks(body.columnId as ColumnId, body.orderedIds)
-        ctx.activity.audit('reordered', 'system', { columnId: body.columnId, orderedIds: body.orderedIds })
-        ctx.activity.log('system', `Reordered tasks in ${body.columnId}`)
-        return Response.json({ ok: true as const })
-      } catch (err) {
-        return Response.json({ error: err instanceof Error ? err.message : String(err) }, { status: 500 })
-      }
-    },
-  }),
-
-  defineRoute({
-    path: '/:taskId/complete',
-    method: 'POST',
-    summary: 'Mark a task as complete',
-    params: taskIdParams,
-    body: completeTaskBody,
-    responses: { 200: okResponse, 400: errorResponse, 500: errorResponse },
-    handler: async (_req, ctx, { params, body }) => {
-      const taskId = params.taskId || body.id
-      if (!taskId) {
-        return Response.json({ error: 'taskId required' }, { status: 400 })
-      }
-      const agent = body.agent || 'system'
-      const summary = body.summary || ''
-      try {
-        const { alreadyComplete } = await reportComplete(taskId, agent, summary, 'rest')
-        if (!alreadyComplete) {
-          ctx.activity.log(agent, `Completed task: ${summary}`, { taskId })
-          indexTask(ctx, taskId).catch(() => {})
-        }
-        return Response.json({ ok: true as const, ...(alreadyComplete ? { alreadyComplete: true as const } : {}) })
-      } catch (err) {
-        return Response.json({ error: err instanceof Error ? err.message : String(err) }, { status: 500 })
-      }
-    },
-  }),
-
-  // Declarative search route — replaces the auto-wired `/search` that
-  // `ctx.search.registerContentType` historically created as a side
-  // effect during activate(). Schemas live in @bakin/core/routing;
-  // handler uses ctx.search at request time.
-  searchRoute({ table: 'tasks', description: 'Full-text search across tasks (title, description, log entries, blocked reasons).' }),
-]
-
 // ─── Plugin definition ───────────────────────────────────────────────────
 
 const tasksPlugin: BakinPlugin = definePlugin({
   id: 'tasks',
   name: 'Tasks',
   version: '2.1.0',
-  routes,
+  routes: tasksRoutes,
 
   settingsSchema: {
     fields: [
